Cleanup sample app remove 3rd party libs
