+++ gradle.properties
@@ -1,30 +1,2 @@
 android.useAndroidX=true
 android.enableJetifier=true
-enableIntune=false
-
-##### Enter MamSdk library maven path (e.g. <MamSdk location>/MamSdk/Library/maven) #####
-mamSdkLibraryMaven=
-
-##### Enter MamSdk library tools path (e.g. <MamSdk location>/MamSdk/Library/tools) #####
-mamSdkLibraryTools=
-
-##### Enter MamSdk library libs path (e.g. <MamSdk location>/MamSdk/Library/libs) #####
-mamSdkLibraryLibs=
-
-##### Enter MamSdk version number (e.g. 20.8.5) #####
-mamSdkVersion=
-
-##### Enter App Package Name (package name has to be globally unique) #####
-appPackageName=com.citrix.mvpntestapp
-
-#### Enter Keystore Path (e.g. <folder location>/mycompany.keystore )  #####
-keyStorePath=
-
-#### Enter Keystore Password #####
-keystorePassword=
-
-#### Enter Key Alias #####
-keyAlias=
-
-#### Enter Key Password #####
-keyPassword=
