+++ keystore.properties
@@ -1,0 +1,11 @@
+#### Enter Keystore Path (e.g. <folder location>/mycompany.keystore )  #####
+keyStorePath=
+
+#### Enter Keystore Password #####
+keystorePassword=
+
+#### Enter Key Alias #####
+keyAlias=
+
+#### Enter Key Password #####
+keyPassword=
