+++ src/main/java/com/citrix/mvpntestapp/activities/WebViewActivity.java
@@ -1,6 +1,8 @@
 package com.citrix.mvpntestapp.activities;
 
 import android.os.Bundle;
+import android.util.Log;
+import android.view.View;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
@@ -11,17 +13,19 @@ import com.citrix.mvpn.api.MicroVPNSDK;
 import com.citrix.mvpn.exception.MvpnException;
 import com.citrix.mvpntestapp.R;
 import com.citrix.mvpntestapp.webview.CustomWebViewClient;
-import com.citrix.sdk.appcore.api.MamSdk;
+
+import static com.citrix.mvpntestapp.activities.MainActivity.URL_KEY;
 
 public class WebViewActivity extends AppCompatActivity {
     private static final String TAG = "Mvpn-WebViewActivity";
+    private WebView webView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.test_webview);
 
-        WebView webView = findViewById(R.id.webview);
+        webView = findViewById(R.id.webview);
         WebViewClient webViewClient = new CustomWebViewClient();
 
         webView.clearCache(true);
@@ -34,9 +38,9 @@ public class WebViewActivity extends AppCompatActivity {
         try {
             webView = MicroVPNSDK.enableWebViewObjectForNetworkTunnel(this, webView, webViewClient);
         } catch (MvpnException e) {
-            MamSdk.getLogger().error(TAG, e.getMessage());
+            Log.e(TAG, e.getMessage());
         }
 
-        webView.loadUrl(getIntent().getStringExtra(SelectStartTunnelActivity.URL_KEY));
+        webView.loadUrl(getIntent().getStringExtra(URL_KEY));
     }
 }
