+++ src/main/java/com/citrix/mvpntestapp/webview/CustomWebViewClient.java
@@ -1,19 +1,20 @@
 package com.citrix.mvpntestapp.webview;
 
 import android.graphics.Bitmap;
+import android.util.Log;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 
-import com.citrix.sdk.appcore.api.MamSdk;
-
 public class CustomWebViewClient extends WebViewClient {
+    private static final String TAG = "CustomWebViewClient";
+
     @Override
     public void onPageStarted(WebView view, String url, Bitmap favicon) {
-        MamSdk.getLogger().debug5("CustomWebViewClient", "WebViewClient onPageStarted:" + url);
+        Log.d(TAG, "WebViewClient onPageStarted:" + url);
     }
 
     @Override
     public void onPageFinished(WebView view, String url) {
-        MamSdk.getLogger().debug5("CustomWebViewClient", "WebViewClient onPageFinished:" + url);
+        Log.d(TAG,"WebViewClient onPageFinished:" + url);
     }
 }
