--- src/main/res/layout/activity_intune_mam_enrollment.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.citrix.mvpntestapp.activities.IntuneMAMEnrollmentActivity">
-
-</androidx.constraintlayout.widget.ConstraintLayout>
