--- src/main/res/layout/activity_select_mam_provider.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:gravity="center"
-    android:orientation="vertical">
-
-    <Button
-        android:id="@+id/btnSecureHub"
-        style="@style/MenuItemButton"
-        android:layout_margin="8dp"
-        android:onClick="onSecureHubClicked"
-        android:text="CEM Secure Hub" />
-
-    <Button
-        android:id="@+id/btnCompanyPortal"
-        style="@style/MenuItemButton"
-        android:layout_margin="8dp"
-        android:onClick="onCompanyPortalClicked"
-        android:text="Intune Company Portal" />
-</LinearLayout>
